Previous timer experiments scrapped. New timer implemented using JS. Now works and saves to database. Case 3 data entry controller refactored and completed.

+++ capstone-practice-track/src/main/java/org/launchcode/capstonepracticetrack/SkillDataRow.java
@@ -13,7 +13,7 @@ public class SkillDataRow {
 
     private ArrayList<HashMap<PracticeSession, Integer>> skillChunks = new ArrayList<>();
 
-    private int total = 0;
+    private double total = 0;
 
     private double average = 0;
 
@@ -88,7 +88,7 @@ public class SkillDataRow {
         this.average = Helpers.round(this.total / this.skillChunks.size(), 2);
     }
 
-    public int getTotal() {
+    public double getTotal() {
         return total;
     }
